clang: remove runtime dependency to perl
Commit https://github.com/kraj/meta-clang/commit/68ec449f97ffa58d in meta-clang adds a dependency to perl,
however at the moment perl does not cross-compile for darwin (the build tries to use readelf which does not
exist on darwin, instead of using objdump).

This dependency is needed only for optional runtime tools, so just remove it at the moment.

Signed-off-by: Etienne Cordonnier <ecordonnier@snap.com>